Se agregaron todos los metodos para actualizar. Tambien se agrego en las clases un metodo que se usara despues

+++ src/main/java/trivia/App.java
@@ -104,22 +104,76 @@ public class App
 	post("/deleteuser", (req, res) -> {
         Map<String, Object> bodyParams = new Gson().fromJson(req.body(), Map.class);
         User u = User.findById(bodyParams.get("id"));
-	    u.delete();
- 		res.type("application/json");
-        return u.toJson(true,"id","dni");
+		if(u!=null){
+	  		u.delete();
+ 			res.type("application/json");
+     		return "Se ha borrado"+u.toJson(true,"id","dni");
+		}
+		return "Error: No se pudo borrar.No se encontraron registro de la persona";
       });
 
 	//actualizar datos de un usuario
 	post("/updateuser", (req, res) -> {
         Map<String, Object> bodyParams = new Gson().fromJson(req.body(), Map.class);
         User u = User.findById(bodyParams.get("id"));
-		u.set("dni", bodyParams.get("dni"));
-		u.set("lastname", bodyParams.get("lastname"));
-		u.set("name", bodyParams.get("name"));
-		u.saveIt();		
+		if(u!=null){
+			u.set("dni", bodyParams.get("dni"));
+			u.set("lastname", bodyParams.get("lastname"));
+			u.set("name", bodyParams.get("name"));
+			u.saveIt();	
+			res.type("application/json");
+			return "Actualizado con exito : "+u.toJson(true);
+		}	
 		res.type("application/json");
-        return u.toJson(true);
+        return "Error: No se pudo actualizar.No se encontraron registros de la persona";
       });
+
+	//actualizar datos de una pregunta
+	post("/updatequestions", (req, res) -> {
+        Map<String, Object> bodyParams = new Gson().fromJson(req.body(), Map.class);
+        Question q = Question.findById(bodyParams.get("id"));
+		if(q!=null){
+			q.set("descrip_q", bodyParams.get("descrip_q"));
+			q.set("id_cat", bodyParams.get("id_cat"));
+			q.saveIt();	
+			res.type("application/json");
+      		return "Actualizado con exito : "+q.toJson(true);
+		}	
+		res.type("application/json");
+        return "Error: No se pudo actualizar.No se encontraron registros de la persona";
+      });
+
+	//actualizar datos de una opcion
+	post("/updateoptions", (req, res) -> {
+        Map<String, Object> bodyParams = new Gson().fromJson(req.body(), Map.class);
+        Option o = Option.findById(bodyParams.get("id"));
+		if(o!=null){
+			o.set("id_q", bodyParams.get("id_q"));
+			o.set("descrip_o", bodyParams.get("descrip_o"));
+			o.set("correct", bodyParams.get("correct"));
+			o.saveIt();	
+			res.type("application/json");
+      		return "Actualizado con exito : "+o.toJson(true);
+		}	
+		res.type("application/json");
+        return "Error: No se pudo actualizar.No se encontraron registros de la persona";
+      });
+
+	//actualizar datos de una categoria
+	post("/updatecategories", (req, res) -> {
+        Map<String, Object> bodyParams = new Gson().fromJson(req.body(), Map.class);
+        Category c = Category.findById(bodyParams.get("id"));
+		if(c!=null){
+			c.set("name_category", bodyParams.get("name_category"));
+			c.saveIt();	
+			res.type("application/json");
+      		return "Actualizado con exito : "+c.toJson(true);
+		}	
+		res.type("application/json");
+        return "Error: No se pudo actualizar.No se encontraron registros de la persona";
+      });
+
+
     }
 
 }
